Start Menu and HUD overhual, fixed a ton of bugs (especially texture uploading), started on muzzleflash, and fixed all map related crashes

+++ src/generic/draw.h
@@ -40,7 +40,7 @@ void Draw_Fill (int x, int y, int w, int h, float r, float g, float b, float a);
 void Draw_FillByColor (int x, int y, int w, int h, float r, float g, float b, float a);
 void Draw_FadeScreen (void);
 void Draw_String (int x, int y, char *str);
-void Draw_ColoredString(int x, int y, char *text, float r, float g, float b, float a, int scale);
+void Draw_ColoredString(int x, int y, char *text, float r, float g, float b, float a, float scale);
 qpic_t *Draw_PicFromWad (char *name);
 qpic_t *Draw_CachePic (char *path);
